Do not use 3D textures larger than the min-max size
defaultImageSize returns 64, 64, 8 for 3D textures, and then
we multiply 64 with 8, and get a width of 320. This is larger than
the minimum value for GL_MAX_3D_TEXTURE_SIZE, which leads to a test
failure on valid implementations.

Decreasing NUM_INVOCATIONS_PER_PIXEL to 4 means we'll end up with
exactly a width of 256, which is guaranteed to be supported. It seems
better to modify this than defaultImageSize as that would change
tests that currently works just fine.

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>