drop check that fails due to NVIDIA driver bug
This test-case is known to be broken, due to a known and reported bug
in the NVIDIA D3D12 driver. This bug is expected to be fixed in the
near future.

The bug in question makes the two 4-component vectors used for the
clip-distances overlap in memory, causing the second set to overwrite
the first set.